Minor efficiency tweaks and QoL changes

+++ fdp.sh
@@ -12,18 +12,31 @@
 #uncomment on CC
 module use /project/6004902/modulefiles
 module load presto
-AFP="/home/adamdong/CHIME-Pulsar_automated_filterbank/"
+
+# check if CHIPSPIPE_DIR is set as an env variable, and use that if possible
+if [[ -z "${CHIPSPIPE_DIR}" ]];then
+    # CHIPSPIPE_DIR isn't set, so use this default
+    AFP="/home/${USER}/CHIME-Pulsar_automated_filterbank"
+else
+    AFP="${CHIPSPIPE_DIR}"
+fi
+
 #comment out on cc
 # SLURM_TMPDIR="/media/adam/1c126a4b-fb16-4471-909f-4b0fda74a5d2/tmpdir"
 SCRATCH_DIR=$(pwd)
 FIL=$1
-cp $FIL ${SLURM_TMPDIR}
-#go to compute node
-cd ${SLURM_TMPDIR}
-python $AFP/fdp.py $FIL
-#come back from compute node
-cd $SCRATCH_DIR
-PULSAR=$(echo "$FIL" | cut -f 1 -d '.')
-cp ${SLURM_TMPDIR}/"$PULSAR"_fdp.fil $SCRATCH_DIR
-#clear tmpdir
-rm ${SLURM_TMPDIR}/*
+
+if [[ -d "$AFP" ]] && [[ -f "$AFP/fdp.py" ]];then
+    cp $FIL ${SLURM_TMPDIR}
+    #go to compute node
+    cd ${SLURM_TMPDIR}
+    python $AFP/fdp.py $FIL
+    #come back from compute node
+    cd $SCRATCH_DIR
+    PULSAR=$(echo "$FIL" | cut -f 1 -d '.')
+    cp ${SLURM_TMPDIR}/"$PULSAR"_fdp.fil $SCRATCH_DIR
+    #clear tmpdir
+    rm ${SLURM_TMPDIR}/*
+else
+    echo "Problem finding AFP directory of AFP/fdp.py file"
+fi
